Worked on deleteCountry functionality
Got the isForeignKey() method working. It finds a record from the repository based on the id you give it, then builds a list of the foreign key column names based off the record retrieved. It then takes the list of repositories and list of fk column names and stores each index in a HashMap before streaming over the HashMap and checking if the record contains a method getter for the fk column. If so, it invokes the getter for the record to get the fk id and tries to look it up in the corresponding repository. If it exists, a DataIntegrityViolationException is thrown and you can't delete the item. If not, you're safe to delete the item from the database.

I renamed the governmentId and continentId fields in Country to facilitate the isForeignKey lookup and to better match the table column names. I also added relevant error messages.

I still need to write tests for more cases but wanted to commit this so I didn't lose my work.

+++ src/main/java/com/mcommings/campaigner/services/CountryService.java
@@ -1,9 +1,7 @@
 package com.mcommings.campaigner.services;
 
 import com.mcommings.campaigner.interfaces.ICountry;
-import com.mcommings.campaigner.models.Continent;
 import com.mcommings.campaigner.models.Country;
-import com.mcommings.campaigner.models.Government;
 import com.mcommings.campaigner.models.RepositoryHelper;
 import com.mcommings.campaigner.models.repositories.IContinentRepository;
 import com.mcommings.campaigner.models.repositories.ICountryRepository;
@@ -57,9 +55,9 @@ public class CountryService implements ICountry {
         if(RepositoryHelper.cannotFindId(countryRepository, countryId)) {
             throw new IllegalArgumentException(DELETE_NOT_FOUND.message);
         }
-//        if(RepositoryHelper.isForeignKey(countryRepository, getListOfRelatedRepositories(), countryId)) {
-//            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
-//        }
+        if(RepositoryHelper.isForeignKey(countryRepository, getListOfRelatedRepositories(), countryId)) {
+            throw new DataIntegrityViolationException(DELETE_FOREIGN_KEY.message);
+        }
         countryRepository.deleteById(countryId);
     }
 
@@ -73,8 +71,8 @@ public class CountryService implements ICountry {
         Country countryToUpdate = RepositoryHelper.getById(countryRepository, countryId);
         countryToUpdate.setName(country.getName());
         countryToUpdate.setDescription(country.getDescription());
-        countryToUpdate.setContinentId(country.getContinentId());
-        countryToUpdate.setGovernmentId(country.getGovernmentId());
+        countryToUpdate.setFk_continent(country.getFk_continent());
+        countryToUpdate.setFk_government(country.getFk_government());
     }
 
     private List<CrudRepository> getListOfRelatedRepositories() {
